+++ src/ladb_opencutlist/js/plugins/jquery.ladb.tab-cutlist.js
@@ -48,6 +48,7 @@
     var SETTING_KEY_CUTTINGDIAGRAM2D_OPTION_SHEET_FOLDING = 'cutlist.cuttingdiagram1d.option.sheet_folding';
     var SETTING_KEY_CUTTINGDIAGRAM2D_OPTION_HIDE_CROSS = 'cutlist.cuttingdiagram2d.option.hide_cross';
     var SETTING_KEY_CUTTINGDIAGRAM2D_OPTION_HIDE_PART_LIST = 'cutlist.cuttingdiagram2d.option.hide_part_list';
+    var SETTING_KEY_CUTTINGDIAGRAM2D_OPTION_ORIGIN_POSITION = 'cutlist.cuttingdiagram2d.option.origin_position';
 
     var EXPORT_DEFAULT_COLUMNS = {
         0 /* EXPORT_OPTION_SOURCE_SUMMARY */        : [ 'material_type', 'material_thickness', 'part_count', 'total_cutting_length', 'total_cutting_area', 'total_cutting_volume', 'total_final_area' ],
@@ -1718,7 +1719,7 @@
                             saw_kerf: that.dialog.getSetting(SETTING_KEY_CUTTINGDIAGRAM1D_OPTION_SAW_KERF + '_' + groupId, that.dialog.getSetting(SETTING_KEY_CUTTINGDIAGRAM1D_OPTION_SAW_KERF, appDefaults.saw_kerf)),
                             trimming: that.dialog.getSetting(SETTING_KEY_CUTTINGDIAGRAM1D_OPTION_TRIMMING + '_' + groupId, that.dialog.getSetting(SETTING_KEY_CUTTINGDIAGRAM1D_OPTION_TRIMMING, appDefaults.trimming)),
                             bar_folding: that.dialog.getSetting(SETTING_KEY_CUTTINGDIAGRAM1D_OPTION_BAR_FOLDING + '_' + groupId, that.dialog.getSetting(SETTING_KEY_CUTTINGDIAGRAM1D_OPTION_BAR_FOLDING, appDefaults.bar)),
-                            hide_cross: that.dialog.getSetting(SETTING_KEY_CUTTINGDIAGRAM1D_OPTION_HIDE_CROSS + '_' + groupId, that.dialog.getSetting(SETTING_KEY_CUTTINGDIAGRAM1D_OPTION_HIDE_CROSS, appDefaults.hide_part_list)),
+                            hide_cross: that.dialog.getSetting(SETTING_KEY_CUTTINGDIAGRAM1D_OPTION_HIDE_CROSS + '_' + groupId, that.dialog.getSetting(SETTING_KEY_CUTTINGDIAGRAM1D_OPTION_HIDE_CROSS, appDefaults.hide_cross)),
                             hide_part_list: that.dialog.getSetting(SETTING_KEY_CUTTINGDIAGRAM1D_OPTION_HIDE_PART_LIST + '_' + groupId, that.dialog.getSetting(SETTING_KEY_CUTTINGDIAGRAM1D_OPTION_HIDE_PART_LIST, appDefaults.hide_part_list)),
                             wrap_length: that.dialog.getSetting(SETTING_KEY_CUTTINGDIAGRAM1D_OPTION_WRAP_LENGTH + '_' + groupId, that.dialog.getSetting(SETTING_KEY_CUTTINGDIAGRAM1D_OPTION_WRAP_LENGTH, appDefaults.wrap_length)),
                         };
@@ -2020,7 +2021,9 @@
                 SETTING_KEY_CUTTINGDIAGRAM2D_OPTION_STACKING,
                 SETTING_KEY_CUTTINGDIAGRAM2D_OPTION_SHEET_FOLDING,
                 SETTING_KEY_CUTTINGDIAGRAM2D_OPTION_HIDE_CROSS,
+                SETTING_KEY_CUTTINGDIAGRAM2D_OPTION_HIDE_CROSS,
                 SETTING_KEY_CUTTINGDIAGRAM2D_OPTION_HIDE_PART_LIST,
+                SETTING_KEY_CUTTINGDIAGRAM2D_OPTION_ORIGIN_POSITION,
 
                 SETTING_KEY_CUTTINGDIAGRAM2D_OPTION_STD_SHEET + '_' + groupId,
                 SETTING_KEY_CUTTINGDIAGRAM2D_OPTION_SCRAP_SHEET_SIZES + '_' + groupId,
@@ -2031,6 +2034,7 @@
                 SETTING_KEY_CUTTINGDIAGRAM2D_OPTION_SHEET_FOLDING + '_' + groupId,
                 SETTING_KEY_CUTTINGDIAGRAM2D_OPTION_HIDE_CROSS + '_' + groupId,
                 SETTING_KEY_CUTTINGDIAGRAM2D_OPTION_HIDE_PART_LIST + '_' + groupId,
+                SETTING_KEY_CUTTINGDIAGRAM2D_OPTION_ORIGIN_POSITION + '_' + groupId,
 
             ],
             2 /* SETTINGS_RW_STRATEGY_MODEL */,
@@ -2043,6 +2047,8 @@
                     } else {
 
                         var appDefaults = response.defaults;
+
+                        console.log(appDefaults);
 
                         var cuttingdiagram2dOptions = {
                             std_sheet: that.dialog.getSetting(SETTING_KEY_CUTTINGDIAGRAM2D_OPTION_STD_SHEET + '_' + groupId, ''),
@@ -2057,6 +2063,7 @@
                             sheet_folding: that.dialog.getSetting(SETTING_KEY_CUTTINGDIAGRAM2D_OPTION_SHEET_FOLDING + '_' + groupId, that.dialog.getSetting(SETTING_KEY_CUTTINGDIAGRAM2D_OPTION_SHEET_FOLDING, appDefaults.sheet_folding)),
                             hide_cross: that.dialog.getSetting(SETTING_KEY_CUTTINGDIAGRAM2D_OPTION_HIDE_CROSS + '_' + groupId, that.dialog.getSetting(SETTING_KEY_CUTTINGDIAGRAM2D_OPTION_HIDE_CROSS, appDefaults.hide_cross)),
                             hide_part_list: that.dialog.getSetting(SETTING_KEY_CUTTINGDIAGRAM2D_OPTION_HIDE_PART_LIST + '_' + groupId, that.dialog.getSetting(SETTING_KEY_CUTTINGDIAGRAM2D_OPTION_HIDE_PART_LIST, appDefaults.hide_part_list)),
+                            origin_position: that.dialog.getSetting(SETTING_KEY_CUTTINGDIAGRAM2D_OPTION_ORIGIN_POSITION + '_' + groupId, that.dialog.getSetting(SETTING_KEY_CUTTINGDIAGRAM2D_OPTION_ORIGIN_POSITION, appDefaults.origin_position)),
                         };
 
                         rubyCallCommand('materials_get_attributes_command', { name: group.material_name }, function (response) {
@@ -2081,6 +2088,7 @@
                             var $selectSheetFolding = $('#ladb_select_sheet_folding', $modal);
                             var $selectHideCross = $('#ladb_select_hide_cross', $modal);
                             var $selectHidePartList = $('#ladb_select_hide_part_list', $modal);
+                            var $selectOriginPosition = $('#ladb_select_origin_position', $modal);
                             var $btnCuttingdiagramOptionsDefaultsSave = $('#ladb_btn_cuttingdiagram_options_defaults_save', $modal);
                             var $btnCuttingdiagramOptionsDefaultsReset = $('#ladb_btn_cuttingdiagram_options_defaults_reset', $modal);
                             var $btnCuttingdiagramOptionsDefaultsResetNative = $('#ladb_btn_cuttingdiagram_options_defaults_reset_native', $modal);
@@ -2115,6 +2123,8 @@
                             $selectHideCross.selectpicker(SELECT_PICKER_OPTIONS);
                             $selectHidePartList.val(cuttingdiagram2dOptions.hide_part_list ? '1' : '0');
                             $selectHidePartList.selectpicker(SELECT_PICKER_OPTIONS);
+                            $selectOriginPosition.val(cuttingdiagram2dOptions.origin_position);
+                            $selectOriginPosition.selectpicker(SELECT_PICKER_OPTIONS);
 
                             var fnEditMaterial = function (callback) {
 
@@ -2153,6 +2163,7 @@
                                 $selectSheetFolding.selectpicker('val', (isAppDefaults ? appDefaults.sheet_folding : that.dialog.getSetting(SETTING_KEY_CUTTINGDIAGRAM2D_OPTION_SHEET_FOLDING, appDefaults.sheet_folding)) ? '1' : '0');
                                 $selectHideCross.selectpicker('val', (isAppDefaults ? appDefaults.hide_cross : that.dialog.getSetting(SETTING_KEY_CUTTINGDIAGRAM2D_OPTION_HIDE_CROSS, appDefaults.hide_cross)) ? '1' : '0');
                                 $selectHidePartList.selectpicker('val', (isAppDefaults ? appDefaults.hide_part_list : that.dialog.getSetting(SETTING_KEY_CUTTINGDIAGRAM2D_OPTION_HIDE_PART_LIST, appDefaults.hide_part_list)) ? '1' : '0');
+                                $selectOriginPosition.selectpicker('val', isAppDefaults ? appDefaults.origin_position : that.dialog.getSetting(SETTING_KEY_CUTTINGDIAGRAM2D_OPTION_ORIGIN_POSITION, appDefaults.origin_position));
                             };
 
                             $inputStdSheet.on('changed.bs.select', function (e) {
@@ -2181,6 +2192,7 @@
                                 var sheet_folding = $selectSheetFolding.val();
                                 var hide_cross = $selectHideCross.val();
                                 var hide_part_list = $selectHidePartList.val();
+                                var origin_position = $selectOriginPosition.val();
 
                                 // Update default cut options for specific type to last used
                                 that.dialog.setSettings([
@@ -2191,6 +2203,7 @@
                                     { key:SETTING_KEY_CUTTINGDIAGRAM2D_OPTION_SHEET_FOLDING, value:sheet_folding },
                                     { key:SETTING_KEY_CUTTINGDIAGRAM2D_OPTION_HIDE_CROSS, value:hide_cross },
                                     { key:SETTING_KEY_CUTTINGDIAGRAM2D_OPTION_HIDE_PART_LIST, value:hide_part_list },
+                                    { key:SETTING_KEY_CUTTINGDIAGRAM2D_OPTION_ORIGIN_POSITION, value:origin_position },
                                 ], 0 /* SETTINGS_RW_STRATEGY_GLOBAL */);
 
                                 that.dialog.notify(i18next.t('tab.cutlist.cuttingdiagram.options_defaults.save_success'), 'success');
@@ -2225,6 +2238,7 @@
                                 cuttingdiagram2dOptions.sheet_folding = $selectSheetFolding.val() === '1';
                                 cuttingdiagram2dOptions.hide_cross = $selectHideCross.val() === '1';
                                 cuttingdiagram2dOptions.hide_part_list = $selectHidePartList.val() === '1';
+                                cuttingdiagram2dOptions.origin_position = $selectOriginPosition.val();
 
                                 // Store options
                                 that.dialog.setSettings([
@@ -2237,6 +2251,7 @@
                                     { key:SETTING_KEY_CUTTINGDIAGRAM2D_OPTION_SHEET_FOLDING + '_' + groupId, value:cuttingdiagram2dOptions.sheet_folding },
                                     { key:SETTING_KEY_CUTTINGDIAGRAM2D_OPTION_HIDE_CROSS + '_' + groupId, value:cuttingdiagram2dOptions.hide_cross },
                                     { key:SETTING_KEY_CUTTINGDIAGRAM2D_OPTION_HIDE_PART_LIST + '_' + groupId, value:cuttingdiagram2dOptions.hide_part_list },
+                                    { key:SETTING_KEY_CUTTINGDIAGRAM2D_OPTION_ORIGIN_POSITION + '_' + groupId, value:cuttingdiagram2dOptions.origin_position },
                                 ], 2 /* SETTINGS_RW_STRATEGY_MODEL */);
 
                                 rubyCallCommand('cutlist_group_cuttingdiagram_2d', $.extend({ group_id: groupId, part_ids: that.selectionGroupId === groupId ? that.selectionPartIds : null }, cuttingdiagram2dOptions, that.generateOptions), function (response) {